Total amount of water abstracted shown as NaN (#2459)
https://eaflood.atlassian.net/browse/WATER-5291

There is an intermittent bug that is occurring when a return has a start meter reading of zero and a zero meter reading. When submitted, occasionally on the "Abstraction return" page the "Total amount of water abstracted" is being shown as `NaN` instead of zero.

Upon investigation it has been found that this issue is being caused when a meter reading is input for a line and then removed. Then a meter reading of zero is entered on a different line. In the session data this creates a reading value of `undefined` for the line that had the value removed. The logic that populates the lines in the `return_submission_lines` table is only checking for `null` readings when it comes to ignoring a line. Therefore when the line is `undefined` it is attempting to perform a calculation on it resulting in `NaN`.

It has also been noticed that the new validation that was added in https://eaflood.atlassian.net/browse/WATER-5192 to prevent all the lines being blank is causing an issue when creating a nil return. This can be fixed by not performing the validation on a nil return as they have no lines.