Stop defaulting the due date in return logs
As part of dynamic due dates WATER-4991, we will no longer be setting the due date of a return log until we have confirmation it has been sent.

This means when we create a return log the due date will be left null. When running the acceptance tests we've spotted we are still setting 'due date' when we seed a record.

Seeding is based on our test helpers so it is there we need to make the change. This is going to break a bunch of tests but we'll deal with them in our subsequent commits.