Add status to notification presenters (#2463)
https://eaflood.atlassian.net/browse/WATER-5280

When we create a notification and save it, we need to set the initial status value to 'pending'.

This was missed in the recent refactor and was producing a bug of an empty status on the notice page.

This change updates the notification presenters to set the status to pending.

We need to leave the existing status in the notify update presenter. This is due to the way we are batch inserting (upserting). If this is not present it will set it to null.