Add preview return form to the communications tab
https://eaflood.atlassian.net/browse/WATER-5295

The legacy code did not save a copy of the return form (paper form) sent to the user. We have updated our notification process to save PDF's sent to Notify, this includes the return form.

We only have return forms as PDF's currently.

This change updates the communications tab to link to the notification page when the notification is a return form.

The notification page has been updated to show a download preview link when the notification PDF has content.

A new service has been added to download the PDF file.