Update Notify Status presenter to handle unknown errors (#2467)
* Update Notify Status presenter to handle unknown errors

https://eaflood.atlassian.net/jira/software/c/projects/WATER/boards/96?selectedIssue=WATER-5280

We found a 'bug' where we were receiving the 'pending-virus-check' response from Notify, but did not map this to letter status. We have added this mapping in this change.

This is not an error but a response indicating that this might take some time and check later, so leaving this as pending means the status update can be checked again.

This also raised a concern where we did not capture unknown statuses from Notify. When this is the case, we will put the raw status into the notification and leave the notification status as 'pending'.