Update PDF batching to 1 per batch
https://eaflood.atlassian.net/browse/WATER-5280

We have some resource constraint on our deployed environments where we are seeing some timeout failures, mostly when creating PDF with Gutenberg.

We could increase the resources in the environments, but we are going to handle this programmatically.

When we batch, we follow the same process for all notifications. As we have built / tested this logic, we do not want to create a separate process for the PDF's. So we will just reduce the batch size to 1.

This means we generate and send a PDF to Notify one at a time.

We have reduced the delay to 2 seconds. This equates to 30 requests a minute way below the constraints imposed by Notify.