Add a new data dates endpoint
One of the known issues with the acceptance tests is that the return invite and reminder journeys rely on you having actual data in your environment.

If you don't then they cannot find due return logs so error. We wanted to fix this as part of stabilising the tests, but to do so the acceptance tests will need to know what the current return periods are.

That's a significant amount of code we don't want to just copy and paste into the acceptance-tests project.

So, we've come up with a data endpoint that returns all the calculated dates the service has to come up with. We know about the current return periods, but there are also the billing periods and current financial year. Should we spot or add any more, we can add them to this endpoint and avoid having to duplicate the code in our test suite.