Fix sort by reference in licence returns
https://eaflood.atlassian.net/browse/WATER-5213

In a [recent change](https://github.com/DEFRA/water-abstraction-system/pull/2346) related to moving to null due dates, we had to update how return logs on the licence returns tab were sorted.

Previously it was by due date, but as these may now be null, we switched to start date and return reference, both descending.

But since then we seem to have ben playing 'test tennis' with our acceptance tests. Someone would run them, and the tests that assert the returns will be in a certain order will fail. That person will fix them, but then someone else reports they fail.

We thought it might be to do with where the local reference number is in relation to the test reference we're applying to the records. It kind of is, but actually it's because we have assumed we are sorting a number, when the column is a varchar.

This change casts the reference to a number for the sort, so that it behaves as expected.