Update licence page to template standard (#2473)
* Update licence page to template standard

https://eaflood.atlassian.net/browse/WATER-5295

We recently updated the communication tab to support the improved return forms.

As we were in this 'space', we need to update the view to the template standard.

This change updates the template to use the 'pageContent' block from the layout.

The notification in this case has HTML styling of strong, so we need to keep this in the view.