Add error capture for return forms PDF generation failure
https://eaflood.atlassian.net/browse/WATER-5280

We are not handling errors when generating a return form very well. They are logged, but it is not necessarily clear what/ when and where we failed during a notification batch.

The legacy code inserts the Notify errors into the notification row when Notify fails. We copied this pattern to be consistent.

This change extends the Notify error column to include errors when Gotenberg (our PDF generation solution) fails. We were seeing timeouts in the test due to lower resources.

We have lifted out the notification error logic into its own presenter so we can consistently insert error into the database.