Migrate search to system
https://eaflood.atlassian.net/browse/WATER-5304

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system app`.

The existing page allows users to search using different types of values, for example, licence references, licence names, and return references.

We are implementing these different types of search into the new ‘system’ page one at a time to manage the complexity.

This change is to implement searching for licences.