Clean up future return cycles during tear down
As part of tearing down the returns schema, this change ensures that any future return cycles are properly cleaned up, preventing residual data from affecting subsequent tests or operations.

When working locally, a dev or QA may add a return version in the future that results in a future return cycle being created. Should this happen, any acceptance test that creates a return version will generate a different number of return logs in that environment.

To ensure the tests pass for all users we need to ensure everyone has the expected return cycles only.