Refactor notice setup returns period
We are adding a new page to allow the user to select the notice types for the standard journey.

We need to go through the existing pages and get them up to 'template standard'. We are doing this so the changes are minimal in the main pr.

This change updates the notice setup returns period logic to 'template standard'.