Housekeeping - Kill flakey assertion
We just want to know the record has been updated and we have that confirmation from the fact the other fields have been overwritten.

The problem is if the test suite runs too fast the creation and update are so close that 'updatedAt' is the same, not greater than the createdAt.