View notice page presenter and view (#2012)
https://eaflood.atlassian.net/browse/WATER-4992

As part of migrating the legacy pages to this app, we have previously replaced the view notices page. Now it's the turn of the view notice page, which is linked to when a user clicks a notice to view.

This PR covers a working page, including filtering (which the legacy version never had).