Update model unit tests (#2478)
Ensure that all the model tests are updated to our new standard, where we just create a one `testRecord` in the DB in a single `before` block. Rather than creating multiple test records in the same test file.

This should help reduce the run time of our unit tests as they will be creating fewer records in the DB.