Void return logs even if no requirements present (#2479)
https://eaflood.atlassian.net/browse/WATER-5302

A user found an issue when making a historic change to a return version. When swapping return cycle type (from summer to all year) it was noticed that we were not voiding the existing return logs for the previous return version. This was due to us exiting the loop early if there were no return requirements to be created for the return cycle.

This PR removes that early exit and allows the void service to be called.