Add standard journey notice types (#2483)
https://eaflood.atlassian.net/browse/WATER-5308

We are updating the 'standard' journey to allow the user to select the notice type. This is replacing the existing two links on the manage page.

A new button will be added in later changes to redirect the use to this page, where they can select the notice type.

The backlink had been added to redirect to the previous page (where the button will be added)