Update FetchNotificationService
The two key changes we make to the service, are selecting more properties and switching from returning the PDF data to just an value that tells us whether we have data or not.

The extra properties will be needed so we can display more information about the notice and the notification in the page.

For the second, we were only returning the PDF data so we could determine whether we could display a link to view the PDF. We weren't trying to display the PDF content itself.

The PDF data could be 50 to 60x greater than the rest of the data combined. So, with some custom SQL, we can have the query do the work, and avoid returning the 100 to 130 KB of data.

We update the tests to covere this, and make use of our now updated fixtures.