Update notice type journey entry point (#2484)
https://eaflood.atlassian.net/browse/WATER-5308

We are updating the 'standard' journey to allow the user to select the notice type. This is replacing the existing two links on the manage page.

This change adds two buttons to the notice page to allow the user to select a standard or adhoc journey.

Navigation has been updated to make sure the 'flow' between pages is as expected.

We no longer provide the notice type to the initiate session, as the user now selects this on select a notice page. (Abstraction alerts do still need this set, but we base our selection on the journey.)

The previous links on the legacy manage page will be hidden behind a feature flag, and this will be the only entry into the notice journey.