Set returned at date on returned letters (#2471)
https://eaflood.atlassian.net/browse/WATER-5296

As part of the work to send out letters to licence holders we need a way to be able to know when a letter has been returned. GovNotify allows you to set up a callback url: https://docs.notifications.service.gov.uk/node.html#returned-letters

When this happens we want to be able to register when we were notified that the letter was returned so this PR sets the returned at date when we receive a request from GovNotify on the callback url.