Update NoticesFixture
We should be using NoticesFixture in conjunction with Notifications filter in our tests. But the NoticesFixture was returning its data in the format of a FetchNoticesService result rather than something we can create an event record from.

With a bit of tweaking we can have it serve both needs. So, we update the raw data so it can be passed to `NoticeHelper.add()` without error. But we also add a function that can map the raw data into the format we expect from FetchNoticesService, and update the existing tests accordingly.

Finally we add a notice type that was not present before and that we may need to test: legacy renewal.