Expand NotificationsFixture with realistic examples
Now we are going to need to test a bunch of notification types, especially when it comes to the presenter.

We want to enhance the data we display, so ideally we want to be working with as many 'realistic' examples as possible.

This update provides 'happy-path' examples for each of the notice types we currently filter by on view notices page (not _every_ abstraction alert type, but we don't need to go _that_ far!)

These are compatible with `NotificationHelper.add()`, or can be amended depending on the needs of the test.

We may enhance, amend and expand them in subsequent commits. But this will help us get started.