Fix broken notice legacy links
https://eaflood.atlassian.net/browse/WATER-5308

When the user is on the manage page, they click a link to start a notice journey. We updated the initiate session to work with buttons from the manage page.

This broke these links. For the time being we need to support multiple entires into the journeys.

This change puts the 'noticeType' back in the initiate session logic.

We are aware the back links will not function as expected. Other changes are being made to make this not an issue.