Set quarterly flag when creating return logs (#2492)
https://eaflood.atlassian.net/browse/WATER-5255

As a lot of reports are driven by the returns table we're adding a column to flag the return as quarterly. This saves us having to join out to the return version table which should optimise the reports that use this table.

This PR updates the return log view and sets the flag when creating the return log.