Replace any use of 'returnForms' with 'paperReturn'
The big switch on.

We previously implemented an adhoc journey for sending a PDF letter for a 'return form'. We used this language throughout the codebase.

This has been updated to 'paper return'. So wherever we referenced 'returnForms', we updated it to be 'paperReturn'.

This changes the notice type from 'returnForms' to 'paperReturn'