Record notice status rather than calculate https://eaflood.atlassian.net/browse/WATER-5306
In this PR we will start recording the overall status of a notice, rather than calculating it. We will also record the count of the different statuses. This removes the need to run a complex query to calculate the status.

There are two points where the status of a notice can change which will be updated.

- After the notice has been ‘sent’ to GOV.UK Notify. A status check is performed to update the status of email notifications. Once complete, the status of the notice will not change until the nightly ‘notification-status’ job runs

- When the nightly-status job runs, it checks the status of notifications with a status of pending. Any notices linked to these notifications will require their status to be recalculated.