Update view presenter to display more
This first pass of the view notification presenter updates it to template standard. But it also extracts more from the notice and notification to display on the page.

We give a bit more about the notice itself, such as the reference code, and who sent it. We extract out who it was sent to from the content.

Not only does this make it easier to see, it means we still show something when there is no 'contents' to display, and it is consistent for all notifications.

Knowing there is very little on the page for paper forms, we extract out details for the return log the form was based on, and provide a link to it.

Taking that same idea, we do the same for abstraction alerts so users don't have to dig into the preview to identify what the monitoring station and threshold was.