Replace any use of 'returnForms' with 'paperReturn' (#2505)
* Refactor 'return-forms' to 'paper-return' for determining and preparing a paper return

We have implemented a 'return-forms' journey for notices on the adhoc journey. This needs to be updated to 'paper-forms'.

This change is for determining and preparing a paper return

This is part of a series of changes to refactor any use of 'return-forms' to 'paper-return'

* Replace any use of 'returnForms' with 'paperReturn'

The big switch on.

We previously implemented an adhoc journey for sending a PDF letter for a 'return form'. We used this language throughout the codebase.

This has been updated to 'paper return'. So wherever we referenced 'returnForms', we updated it to be 'paperReturn'.

This changes the notice type from 'returnForms' to 'paperReturn'