Update notice setup notice type page to show notice type based on scope (#2509)
* Update notice setup notice type page to show notice type based on permissions

https://eaflood.atlassian.net/browse/WATER-5323

This also covers https://eaflood.atlassian.net/browse/WATER-5325.

We recently created a shared page to allow the user to select a notice type. This works for both adhoc and the standard journey.

Logic / conditions have been added to show various Notice types depending on the journey and permissions.

We have temporality used the invitations Notify templates for the ad hoc reminders.