Add error details to the page
We have to handle a few scenarios when it comes to errors. But essentially it breaks down to

- if notify error is populated something went wrong when we tried to send it to Notify
- if notify error is null but status is 'error', then it must be because the notify status indicates a failure.

We can control the description in the second scenario because we take it straight from Notify. But when its a system error we are dealing with thousands of records and we haven't gone digging into what they may contain. Hence we stick with a generic error message.