Handle cancelled status from Notify
https://eaflood.atlassian.net/browse/WATER-5328

When working on [Updating view notification to support all statuses](https://github.com/DEFRA/water-abstraction-system/pull/2507) (not just SENT ones), we came across a subset of almost 3,000 notifications.

These have a status of `SENT` but a [Notify](https://www.notifications.service.gov.uk/) status of `cancelled`. This situation only applies to letters because Notify does not send them to the provider until the evening. This means users can cancel any letter via the Notify dashboard before it is sent.

At the moment, we are displaying these as `SENT`, which is not correct.

In [Record notice status rather than calculate](https://github.com/DEFRA/water-abstraction-system/pull/2506) we updated what happens when we have finished fetching the statuses for a notice (or multiple as part of the 'notification-status' job).

The updated `UpdateEventService` now runs a query that determines the overall status for a notice (`water.event`) as well as getting a count of all the different statues.

With this recent finding, it needs to be updated to take into account this new 'cancelled' status.

But in order for a notification to have a status of 'cancelled', we need to ensure we are setting its status to then when the Notify status response is cancelled.