Update service to handle cancelled plus 1st test
We update the query in the service to record cancelled counts, and determine the overall status based on it.

Our first test demonstrates that when we have a notice with both a sent and a cancelled notification, we set the overall status to `sent` _not_ `cancelled`.

This is how `cancelled` differs from the other statuses. It is the result of an _intended_ action by the user so we don't consider it an issue, and only highlight it if all notifications in the notice were cancelled.