Update view notices with overall_status & cancelled
https://eaflood.atlassian.net/browse/WATER-5331

With the change [Handle cancelled status from Notify](https://github.com/DEFRA/water-abstraction-system/pull/2514) we now have a new status to handle in the view notices page.

This built on [Record notice status (against event) rather than calculate](https://github.com/DEFRA/water-abstraction-system/pull/2506).

We need to be able to display the `CANCELLED` status for a notice, which we are now saving against the notice in `overall_status`.

So, the first change we make is to stop calculating the status and just pull it from `overall_status`. Then we ensure `CANCELLED` is part of our notice status tag macro.