Handle cancelled returned status view notice page (#2516)
- https://eaflood.atlassian.net/browse/WATER-5311
- https://eaflood.atlassian.net/browse/WATER-5333

In WATER-5296 we made a change to the service to start setting the status of letter notifications that were returned to Notify as `RETURNED`.

Then we realised we also needed to handle letter notifications where a user cancels the letter before it is sent for printing in the Notify dashboard. So, in WATER-5328 we [Handle cancelled status from Notify](https://github.com/DEFRA/water-abstraction-system/pull/2514).

We are still working on the view notice page, but with these two new statuses, we need to ensure that we cannot only display them but also filter by them.

To help with that, we also updated notices so that the [overall status is persisted against the event rather than calculated](https://github.com/DEFRA/water-abstraction-system/pull/2506) each time.

So, we can also simplify how we determine the top-level status in the page.