Update view notice metadata to be consistent (#2517)
https://eaflood.atlassian.net/browse/WATER-4992

As part of migrating the legacy pages to this app, we have been working on a new view notice page, which is linked to when a user clicks a notice in the view notices page.

The [initial version was added](https://github.com/DEFRA/water-abstraction-system/pull/2012), but we've since realised that there are some inconsistencies.

- How the metadata is displayed is inconsistent with other pages that contain metadata
- We used created by and created date, instead of sent by and sent date

This change updates the page to be consistent with the rest of the service.