Update notice overall status when letter returned
https://eaflood.atlassian.net/browse/WATER-5335

Now, when Notify receives a returned letter, they will send us a callback. This is a request to the service that tells us a letter notification has been returned, for example, 'Not at this address'.

In [Set returned at date on returned letters](https://github.com/DEFRA/water-abstraction-system/pull/2471) we also set the status of the notification to `RETURNED`.

By updating the status, users should more easily find these returned notifications.

Shortly after this, we started [recording notice status rather than calculate it](https://github.com/DEFRA/water-abstraction-system/pull/2506). Before the change, it was calculated for _every_ notice _every_ time the page was loaded. We were already starting to see it strain, and are soon to add another filter to it.

So, we switched to calculating it at specific points and storing it against the notice as `overall_status`. Those points were.

- When a notice is sent, we check the Notify status of all emails included in the notice.
- When the overnight 'notification-status' job runs. This looks for any 'pending' notifications and checks their status with Notify.

At both times, because the status of the notifications may change, we need to recalculate the `overall_status` and `status_counts` for the linked notice.

What we have overlooked is that when we update the status of a notification to `RETURNED` the same logic applies. We should be recalculating the `overall_status` and `status_counts` for the linked notice.

This change updates the returned letter process to ensure this happens.