Route the standard notice journey through the check notice type page
The 'adhoc' journey has a check page before going through to the check recipients page.

The standard journey also needs this pre check page.

This change adds the logic to use the existing check page for the standard journey.

When showing the standard journey, we want to show the returns period selected with the option of a change link.