Update service to do all work and follow convention
Ideally, controllers should do as little of the work as possible. So, we move everything the controler was doing into the process returned letter service.

This includes logging we got a returned letter, but we update it to use our existing conventions for logging work done.

We'll record the success along with the time taken, plus any pertinent details, in this case the payload from notify an which notification, if any was updated.

We also wrap it in a try catch and handle the error. The legacy notify receipt callback was notorious for bringing down the service when it threw an error and we don't want to make the same mistake.

To support all this, and make whats going on a little clearer, we restructure the tests. It means we are interacting with the DB more, but the clearer tests win out.