Update view notification to support all statuses (#2507)
https://eaflood.atlassian.net/browse/WATER-5327

In support of dynamic due dates, we have implemented the ability to capture when a letter is returned via a [callback from GOV.UK Notify](https://docs.notifications.service.gov.uk/rest-api.html#returned-letters). (See WATER-5296)

This means we’ll know which notification letters were returned, and when we are notified of this. We’ll be changing the status of the notification to `RETURNED`, and users will be able to find them in the view notices page.

But they can’t see the notification that was returned, nor can they see when we were notified.

A similar issue applies to `ERRORED` notifications. We let users know the notification is ‘errored’, but first, they have to find it in the view notices page, assuming they know which notice it was sent in. And we don’t really tell them much else.

We have a view notification page, but it is only accessible from the view licence communications page, and it only lists out `SENT` notifications.

This change updates the view notification page to handle displaying _any_ notification, regardless of its state.

It includes displaying extra details about the notification, for example, why a notification has been marked as errored. We also display context for specific notices, such as the monitoring station, restriction, and flow for an abstraction alert. Or details from the return for a paper return.