Remove redundant notices routes from Manage page
https://eaflood.atlassian.net/browse/WATER-5207

We knew that some of the links in the manage page led to features that were either no longer used, or should not be used.

These were

- Send flow notices
  - Restriction
  - Hands-off flow
  - Resume
- Send line notices
  - Renewal

With us taking over the notice journeys, and about to release the last of them (ad-hoc replacing the legacy paper returns) we really don't want the legacy code messing up all the data we've just fixed!

Before we do this we have checked the stats. The send flow notice routes were only used twice since 2025-04-01 (time of writing it is 2025-10-13). Renewal has not been used since May 2023!

So, before we ship our version of the 'Manage' page we're dropping these links from it.