Link view notice page to view notification page
https://eaflood.atlassian.net/browse/WATER-4992

In [View notice page presenter and view](https://github.com/DEFRA/water-abstraction-system/pull/2012) we added our first iteration of a replacement for the legacy view notice page.

What we really wanted to do was allow users to link from there to viewing a notification. But the page as it stood had been built only to be accessed from the view licence communications page, and it only shows notifications with a status of `SENT`.

We [updated the page](https://github.com/DEFRA/water-abstraction-system/pull/2507) so it will now handle _all_ notifications in any state. Now we need to update the view notice page to link to it.