Link view notice page to view notification page (#2525)
https://eaflood.atlassian.net/browse/WATER-4992

In [View notice page presenter and view](https://github.com/DEFRA/water-abstraction-system/pull/2012), we added our first iteration of a replacement for the legacy view notice page.

What we really wanted to do was allow users to link from there to viewing a notification. But the page, as it stood, had been built only to be accessed from the view licence communications page, and it only shows notifications with a status of `SENT`.

We [updated the page](https://github.com/DEFRA/water-abstraction-system/pull/2507) so it will now handle _all_ notifications in any state. Now we need to update the view notice page to link to it.

A notification can be linked to multiple licences. When we come from the view notice page, we won't be able to provide a licence ID. So, the other change we make is to update the view notification page based on the context.

When the licence ID is provided.

- The active search bar has 'search' highlighted
- The back link goes to the licence communications page
- The title caption is 'Licence 01/123'
- The notice's reference code is displayed in the metadata

When the licence ID is _not_ provided.

- The active search bar has 'manage' highlighted
- The back link goes to the notice page
- The title caption is 'Notice RINV-1234'
- The notice's reference code is _not_ displayed in the metadata (we've displayed twice already!)