Remove view-notice-feature-flag
https://eaflood.atlassian.net/browse/WATER-4992

When we started working on a replacement view notice page, we added a feature flag as we normally do so we have the option of reverting to the legacy if there is an issue.

However, with all the changes we have been doing related to notices and notifications, we are seeing the legacy page constantly fall over because something in either the notice or a notification is not as expected. Whereas the new view notice page is not having any of these issues.

So, we have decided to ditch the feature flag and just make the new page the default. If there are issues found we'll fix forward rather than wasting time trying to make the legacy version work.