Default overall status and counts when sending (#2528)
https://eaflood.atlassian.net/browse/WATER-5306

In [Record notice status rather than calculate](https://github.com/DEFRA/water-abstraction-system/pull/2506), we changed how we determine the status to display for a notice in the view notices and view notice pages.

Before, we were always calculating the status. But we were experiencing a performance issue, and we were asked to add a filter to view notices. This will be much easier if the overall status is already determined.

So, the change meant that after sending a notice, when we check its status in the overnight job, and if we get a returned letter callback from Notify, we run `UpdateEventService`, which gathers the counts of the notifications linked to the notice. From this, it determines an overall status and then applies it to the notice (`water.events`).

This is working great, apart from one issue. When sending paper returns, which require generating and uploading PDFs to Notify, the process can take a few seconds.

In this time, a user is free to go to the view notice page to check how things are going. Because `UpdateEventService` has not yet run (it's the last step), the overall status is borked. We get a tiny blue square, which tells us we've not set a status.

To resolve this, we're going to start setting `overall_status` and `status_counts` to default values so we don't break anything here or in the view notices page.