Remove enable-system-notice-view feature flag (#2529)
https://eaflood.atlassian.net/browse/WATER-4992

When we started working on a replacement view notice page, we added a feature flag as we usually do, so we have the option of reverting to the legacy if there is an issue.

However, with all the changes we have been making related to notices and notifications, we are seeing the legacy page frequently crash because something in either the notice or a notification is not as expected. In comparison, the new view notice page does not have any of these issues.

So, we have decided to ditch the feature flag and make the new page the default. If there are issues found, we'll fix-forward rather than wasting time trying to make the legacy version work.